+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## July, 2025
+### 23/07
+- Now it's possible to take a full screenshot with Print Screen, or crop a region yourself with Ctrl + Print Screen, images are temporarily going into the root directory.
 ### 22/07
 - Project rebuilt on .NET 8 (previous was .Net Framework 4.7.2)
 - SQLite package installed